# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Competency Platform REST API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /evaluations:
    get:
      operationId: listEvaluations
      tags: [evaluations]
      parameters:
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: Evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evaluation"
    post:
      operationId: createEvaluation
      tags: [evaluations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /evaluations/form-data:
    get:
      operationId: evaluationFormData
      tags: [evaluations]
      parameters:
        - { name: employeeId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationFormData"
  /evaluations/department-distribution:
    get:
      operationId: departmentDistribution
      tags: [evaluations]
      parameters:
        - { name: employeeId, in: query, required: true, schema: { type: string } }
        - { name: period, in: query, schema: { type: string } }
        - { name: excludeEvaluationId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDistribution"
  /evaluations/{id}:
    get:
      operationId: getEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
    put:
      operationId: updateEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /evaluations/{id}/submit:
    post:
      operationId: submitEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationStatusResponse"
  /evaluations/{id}/approve:
    post:
      operationId: approveEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationStatusResponse"
  /evaluations/{id}/reject:
    post:
      operationId: rejectEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationStatusResponse"
  /evaluations/{id}/acknowledge:
    post:
      operationId: acknowledgeEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationStatusResponse"
  /evaluations/{id}/object:
    post:
      operationId: objectEvaluation
      tags: [evaluations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectionInput"
      responses:
        "200":
          description: Objected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationStatusResponse"
...
components:
  schemas:
...
    IdResponse:
      type: object
      properties:
        id: { type: string }
      required: [id]
...
    EvaluationStatusResponse:
      type: object
      properties:
        id: { type: string }
        status: { type: string }
      required: [id, status]
...
    EvaluationItem:
      type: object
      properties:
        id: { type: string }
        evaluationId: { type: string }
        kind: { type: string }
        refKey: { type: string }
        label: { type: string }
        score: { type: ["number", "null"] }
        note: { type: ["string", "null"] }
      required: [id, kind, refKey, label]
    Evaluation:
      type: object
      properties:
        id: { type: string }
        employeeId: { type: string }
        jobId: { type: ["string", "null"] }
        evaluatorId: { type: ["string", "null"] }
        approverId: { type: ["string", "null"] }
        period: { type: string }
        mode: { type: string }
        kpiWeight: { type: integer }
        competencyWeight: { type: integer }
        kpiScore: { type: ["number", "null"] }
        competencyScore: { type: ["number", "null"] }
        totalScore: { type: ["number", "null"] }
        ratingLabel: { type: ["string", "null"] }
        status: { type: string }
        approvedAt: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        employeeAck: { type: boolean }
        objectionNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationItem"
        employeeName: { type: ["string", "null"] }
        evaluatorName: { type: ["string", "null"] }
      required: [id, employeeId, period, mode, status]
    EvaluationKpiInput:
      type: object
      properties:
        name: { type: string }
        achievement: { type: number }
        note: { type: ["string", "null"] }
      required: [name, achievement]
    EvaluationInput:
      type: object
      properties:
        employeeId: { type: string }
        period: { type: string }
        mode: { type: string }
        kpiWeight: { type: integer }
        sharedScores:
          type: object
          additionalProperties: { type: number }
        jobScores:
          type: object
          additionalProperties: { type: number }
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationKpiInput"
      required: [employeeId, period, mode]
    EvaluationUpdate:
      type: object
      properties:
        period: { type: string }
        mode: { type: string }
        kpiWeight: { type: integer }
        sharedScores:
          type: object
          additionalProperties: { type: number }
        jobScores:
          type: object
          additionalProperties: { type: number }
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/EvaluationKpiInput"
      required: [period, mode]
    RejectInput:
      type: object
      properties:
        reason: { type: string }
      required: [reason]
    ObjectionItemInput:
      type: object
      properties:
        itemId: { type: string }
        note: { type: ["string", "null"] }
      required: [itemId]
    ObjectionInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ObjectionItemInput"
      required: [items]
    FormEmployee:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        employeeNumber: { type: string }
        jobId: { type: ["string", "null"] }
        jobName: { type: ["string", "null"] }
        departmentName: { type: ["string", "null"] }
      required: [id, name, employeeNumber]
    FormSharedItem:
      type: object
      properties:
        key: { type: string }
        name: { type: string }
        type: { type: string }
      required: [key, name, type]
    FormJobCompetency:
      type: object
      properties:
        key: { type: string }
        name: { type: string }
        type: { type: string }
      required: [key, name, type]
    FormKpi:
      type: object
      properties:
        name: { type: string }
        target: { type: ["string", "null"] }
        weight: { type: ["number", "null"] }
      required: [name]
    EvaluationFormData:
      type: object
      properties:
        employee:
          $ref: "#/components/schemas/FormEmployee"
        shared:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/FormSharedItem"
        jobCompetencies:
          type: array
          items:
            $ref: "#/components/schemas/FormJobCompetency"
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/FormKpi"
      required: [employee, shared, jobCompetencies, kpis]
    PolicySet:
      type: object
      properties:
        labels:
          type: array
          items: { type: string }
        above:
          type: array
          items: { type: number }
        achieved:
          type: array
          items: { type: number }
        below:
          type: array
          items: { type: number }
      required: [labels, above, achieved, below]
    DepartmentDistribution:
      type: object
      properties:
        departmentName: { type: ["string", "null"] }
        employeeCount: { type: integer }
        counts:
          type: array
          items: { type: integer }
        evaluatedCount: { type: integer }
        achievement: { type: number }
        policy:
          $ref: "#/components/schemas/PolicySet"
      required: [employeeCount, counts, evaluatedCount, achievement, policy]
...